Insanity but less so, merging is somewhat successful
