+++ Scripts/5DGameLogic/5DGameEngine/Move.cs
@@ -1,15 +1,16 @@
 using System;
+using FileIO5D;
 
-namespace Engine
+namespace FiveDChess
 {
 	public class Move : IComparable<Move>
 	{
-		//TODO Make this make sense more, for example extend this for promotion or caslting?
-		public CoordFive Origin { get; set; }
-		public CoordFive Dest { get; set; }
-		public int PieceMoved { get; set; }
-		// 1 for spatial, 2 for jumping(unrecognized branching, 3 for branching.
-		public int Type { get; set; }
+		public CoordFive Origin;
+		public CoordFive Dest;
+		public int PieceMoved;
+		// 1 for spatial, 2 for jumping(unrecognized branching), 3 for branching.
+		public int Type;
+		public int SpecialType;
 
 		public static readonly int NULLMOVE = -1;
 		public static readonly int NORMALMOVE = 0;
@@ -18,10 +19,9 @@ namespace Engine
 		public static readonly int SPATIALMOVE = 1;
 		public static readonly int JUMPINGMOVE = 2;
 		public static readonly int BRANCHINGMOVE = 3;
-
+		public static readonly Move NULL = new Move(new CoordFive(0,0,0,0,true),new CoordFive(0,0,0,0,true));
 		// 0 is normal, 1 is castling, and 2+ is promotion with the type being promoted.
 		// For castling, origin in the king and dest is the rook.
-		public int SpecialType { get; set; }
 
 		public Move(CoordFive coordOrigin, CoordFive coordDest)
 		{
@@ -58,14 +58,24 @@ namespace Engine
 		public Move(CoordFive boardOrigin)
 		{
 			Origin = boardOrigin;
-			Dest = new CoordFive(0, 0, 0, 0,boardOrigin.Color);//ENSURE this works correctly with color added.
+			Dest = new CoordFive(0, 0, 0, 0,boardOrigin.Color);
 			Type = SPATIALMOVE;
 			SpecialType = NULLMOVE;
 		}
 
-		public string RawMoveNotation()
+		public Move Clone()
 		{
-			return Origin.RawCoordString() + Dest.RawCoordString();
+			Move newMove = new Move(Origin.Clone(),Dest.Clone(),SpecialType);
+			newMove.Type = Type;
+			newMove.PieceMoved = PieceMoved;
+			newMove.SpecialType = SpecialType;
+			return newMove;
+		}
+
+		public void SwapColor()
+		{
+			Origin.Color = !Origin.Color;
+			Dest.Color = !Dest.Color;
 		}
 
 		public override string ToString()
@@ -83,45 +93,17 @@ namespace Engine
 			string move = "";
 			if (Type == SPATIALMOVE)
 			{
-				move += "(" + Origin.L + "T" + Origin.T + ")" + piece + Dest.SANString();
+				move += "(" + Origin.L + "T" + Origin.T + ")" + piece + StringUtils.SANString(Dest);
 			}
 			else if (Type == JUMPINGMOVE)
 			{
-				move += "(" + Origin.L + "T" + Origin.T + ")" + piece + Origin.SANString() + ">("
-						+ Dest.L + "T" + Dest.T + ")" + Dest.SANString();
+				move += "(" + Origin.L + "T" + Origin.T + ")" + piece + StringUtils.SANString(Origin) + ">("
+						+ Dest.L + "T" + Dest.T + ")" + StringUtils.SANString(Dest);
 			}
 			else if (Type == BRANCHINGMOVE)
 			{
-				move += "(" + Origin.L + "T" + Origin.T + ")" + piece + Origin.SANString() + ">>("
-						+ Dest.L + "T" + Dest.T + ")" + Dest.SANString();
-			}
-			return move;
-		}
-
-		public string ToShadString()
-		{
-			string move = "";
-			int piece = PieceMoved;
-			if (piece > Board.numTypes)
-			{
-				piece -= Board.numTypes;
-			}
-			if (PieceMoved != Board.EMPTYSQUARE && piece != 1)
-			{
-				move += Board.PieceChars[piece];
-			}
-			move = "(" + Origin.L + "T" + Origin.T + ")" + move;
-			if (Type == SPATIALMOVE)
-			{
-				move += Dest.SANString();
-			}
-			else if (Type == JUMPINGMOVE)
-			{
-				move += Origin.SANString() + ">(" + Dest.L + "T" + Dest.T + ")" + Dest.SANString();
-			}
-			else if (Type == BRANCHINGMOVE)
-			{
-				move += ">>(" + Dest.L + "T" + Dest.T + ")" + Dest.SANString();
+				move += "(" + Origin.L + "T" + Origin.T + ")" + piece + StringUtils.SANString(Origin) + ">>("
+						+ Dest.L + "T" + Dest.T + ")" + StringUtils.SANString(Dest);
 			}
 			return move;
 		}
@@ -142,58 +124,6 @@ namespace Engine
 		public bool Equals(Move m)
 		{
 			return this.Origin.Equals(m.Origin) && this.Dest.Equals(m.Dest);
-		}
-
-		public string ToRawShadString()
-		{
-			if (this.SpecialType == CASTLE)
-			{
-				if (Dest.X > Origin.X)
-				{
-					return "O-O";
-				}
-				else
-				{
-					return "O-O-O";
-				}
-			}
-			string move = "";
-			string temporalOrigin = "(" + Origin.L + "T" + Origin.T + ")";
-			string sanOrigin = Origin.SANString();
-			string destStr = "(" + Dest.L + "T" + Dest.T + ")" + Dest.SANString();
-			int piece = this.PieceMoved;
-			if (piece > Board.numTypes)
-			{
-				piece -= Board.numTypes;
-			}
-			move = temporalOrigin;
-			if (PieceMoved != Board.EMPTYSQUARE && piece != 1)
-			{
-				move += Board.PieceChars[piece];
-			}
-			move += sanOrigin;
-			if (Type == SPATIALMOVE)
-			{
-				move += Dest.SANString();
-			}
-			else if (Type == JUMPINGMOVE)
-			{
-				move += ">" + destStr;
-			}
-			else if (Type == BRANCHINGMOVE)
-			{
-				move += ">>" + destStr;
-			}
-			if (this.SpecialType > CASTLE)
-			{
-				int promoted = SpecialType;
-				if (SpecialType > Board.numTypes)
-				{
-					promoted -= Board.numTypes;
-				}
-				move += "=" + Board.PieceChars[promoted];
-			}
-			return move;
 		}
 	}
 }
